[FE Header index][SQL] - Tự Hiếu, Sửa icon cart, bỏ side-menu giỏ hàng, sửa SQL end-date price thành null

+++ laptech/bin/src/main/resources/static/admin/assets/js/angularjs/category.js
@@ -1,6 +1,6 @@
+import { toastMixin, confirmationDialog } from "../global/custom-sweetalert.js";
 let host = "http://localhost:8081/api";
-var i = 0;
-const app = angular.module("app", []);
+// const app = angular.module("app", []);
 app
   .controller("list", list)
   .controller("create", formCreate)
@@ -24,6 +24,7 @@ function list($scope, $http) {
     })
       .then((resp) => {
         $scope.items = resp.data;
+        window.sessionStorage.setItem("items", JSON.stringify(resp.data));
         $scope.pageCount = Math.ceil($scope.items.length / 5);
         console.log("Success1", resp);
       })
@@ -40,7 +41,6 @@ function list($scope, $http) {
       $scope.sortColumn === prop ? !$scope.reverseSort : false;
     $scope.sortColumn = prop;
   };
-
   $scope.begin = 0;
   $scope.pageCount = Math.ceil($scope.items.length / 5);
   $scope.first = function () {
@@ -65,29 +65,44 @@ function list($scope, $http) {
     window.sessionStorage.setItem("editId", id);
   };
   $scope.delete = (id) => {
-    var url = `${host}/category/${id}`;
-    //var url = host+'/students.json';
-    $http({
-      method: "delete",
-      url: url,
-    })
-      .then((resp) => {
-        var index = $scope.items.findIndex((item) => item.id == id);
-        $scope.items.splice(index, 1);
-        $scope.reset();
-        console.log("Success", resp);
-        $scope.pageCount = Math.ceil($scope.items.length / 5);
-      })
-      .catch((error) => {
-        console.log("Error", error);
-        if (error.status === 422) {
-          alert(
-            "Danh mục đã tồn tại trong sản phẩm. Cập nhật không thành công."
-          );
-        } else {
-          alert("Đã xảy ra lỗi khi cập nhật danh mục.");
-        }
-      });
+    confirmationDialog(
+      "Xác nhận xóa?",
+      "Bạn có chắc chắn muốn xóa dữ liệu?",
+      "warning",
+      "Xóa",
+      "Hủy"
+    ).then((result) => {
+      if (result.isConfirmed) {
+        var url = `${host}/category/${id}`;
+        $http({
+          method: "delete",
+          url: url,
+        })
+          .then((resp) => {
+            var index = $scope.items.findIndex((item) => item.id == id);
+            $scope.items.splice(index, 1);
+            $scope.reset();
+            console.log("Success", resp);
+            $scope.pageCount = Math.ceil($scope.items.length / 5);
+            toastMixin.fire({
+              animation: true,
+              icon: "success",
+              title: "Xóa thành công",
+            });
+          })
+          .catch((error) => {
+            console.log("Error", error);
+            toastMixin.fire({
+              animation: true,
+              icon: "error",
+              title:
+                "Danh mục đã tồn tại trong sản phẩm. Cập nhật không thành công.",
+              position: "top",
+              width: 600,
+            });
+          });
+      }
+    });
   };
   //
   $scope.search = (name) => {
@@ -109,34 +124,80 @@ function list($scope, $http) {
         console.log("Error_edit", error);
       });
   };
+
+  $scope.check = () => {
+    const name = window.sessionStorage.getItem("name");
+
+    if (name == "create") {
+      toastMixin.fire({
+        animation: true,
+        icon: "success",
+        title: "Thêm thành công",
+      });
+      window.sessionStorage.removeItem("name");
+    } else if (name == "update") {
+      toastMixin.fire({
+        animation: true,
+        icon: "success",
+        title: "Sửa thành công",
+      });
+      window.sessionStorage.removeItem("name");
+    } else if (name == "delete") {
+      toastMixin.fire({
+        animation: true,
+        icon: "success",
+        title: "Xóa thành công",
+      });
+      window.sessionStorage.removeItem("name");
+    }
+  };
+
   // thực hiện
   $scope.reset();
   $scope.load_all();
+  $scope.check();
 }
 function formCreate($scope, $http) {
   //
+
   $scope.form = {};
   $scope.items = [];
   $scope.create = () => {
-    var item = angular.copy($scope.form);
-    var url = `${host}/category`;
-    $http({
-      method: "post",
-      url: url,
-      data: item,
-    })
-      .then((resp) => {
-        $scope.items.push(item);
-        console.log("Success", resp);
-        window.location.href = "/admin/category/list";
-      })
-      .catch((error) => {
-        console.log("Error", error);
-      });
+    if (validation($scope, $scope.form)) {
+      if (validationCreate($scope.form)) {
+        confirmationDialog(
+          "Xác nhận thêm?",
+          "Bạn có chắc chắn muốn thêm dữ liệu?",
+          "question",
+          "Thêm",
+          "Hủy"
+        ).then((result) => {
+          if (result.isConfirmed) {
+            var item = angular.copy($scope.form);
+            var url = `${host}/category`;
+            $http({
+              method: "post",
+              url: url,
+              data: item,
+            })
+              .then((resp) => {
+                $scope.items.push(item);
+                console.log("Success", resp);
+                window.location.href = "/admin/category/list";
+                window.sessionStorage.setItem("name", "create");
+              })
+              .catch((error) => {
+                console.log("Error", error);
+              });
+          }
+        });
+      }
+    }
   };
 }
 function formUpdate($scope, $http) {
   //
+  $scope.isLoading = true;
   $scope.form = {};
   $scope.items = [];
   $scope.edit = (id) => {
@@ -148,54 +209,85 @@ function formUpdate($scope, $http) {
       .then((resp) => {
         $scope.form = resp.data;
         console.log("Success_edit", resp);
+        $scope.isLoading = false;
       })
       .catch((error) => {
         console.log("Error_edit", error);
       });
     window.sessionStorage.removeItem("editId");
   };
+
   $scope.update = () => {
-    var item = angular.copy($scope.form);
-    var url = `${host}/category/${$scope.form.id}`;
-    $http({
-      method: "put",
-      url: url,
-      data: item,
-    })
-      .then((resp) => {
-        var index = $scope.items.findIndex((item) => item.id == $scope.form.id);
-        $scope.items[index] = resp.data;
-        console.log("Success", resp);
-        window.location.href = "/admin/category/list";
-      })
-      .catch((error) => {
-        console.log("Error", error);
-        if (error.status === 422) {
-          alert(
-            "Danh mục đã tồn tại trong sản phẩm. Cập nhật không thành công."
-          );
-        } else {
-          alert("Đã xảy ra lỗi khi cập nhật danh mục.");
+    if (validation($scope, $scope.form)) {
+      confirmationDialog(
+        "Xác nhận sửa?",
+        "Bạn có chắc chắn muốn sửa dữ liệu?",
+        "question",
+        "Sửa",
+        "Hủy"
+      ).then((result) => {
+        if (result.isConfirmed) {
+          var item = angular.copy($scope.form);
+          var url = `${host}/category/${$scope.form.id}`;
+          $http({
+            method: "put",
+            url: url,
+            data: item,
+          })
+            .then((resp) => {
+              var index = $scope.items.findIndex(
+                (item) => item.id == $scope.form.id
+              );
+              $scope.items[index] = resp.data;
+              console.log("Success", resp);
+              window.location.href = "/admin/category/list";
+              window.sessionStorage.setItem("name", "update");
+            })
+            .catch((error) => {
+              console.log("Error", error);
+            });
         }
       });
+    }
   };
   $scope.delete = (id) => {
-    var url = `${host}/category/${id}`;
-    //var url = host+'/students.json';
-    $http({
-      method: "delete",
-      url: url,
-    })
-      .then((resp) => {
-        var index = $scope.items.findIndex((item) => item.id == $scope.form.id);
-        $scope.items.splice(index, 1);
-        $scope.form = {};
-        console.log("Success", resp);
-        window.location.href = "/admin/category/list";
-      })
-      .catch((error) => {
-        console.log("Error", error);
-      });
+    confirmationDialog(
+      "Xác nhận xóa?",
+      "Bạn có chắc chắn muốn xóa dữ liệu?",
+      "warning",
+      "Xóa",
+      "Hủy"
+    ).then((result) => {
+      if (result.isConfirmed) {
+        var url = `${host}/category/${id}`;
+        //var url = host+'/students.json';
+        $http({
+          method: "delete",
+          url: url,
+        })
+          .then((resp) => {
+            var index = $scope.items.findIndex(
+              (item) => item.id == $scope.form.id
+            );
+            $scope.items.splice(index, 1);
+            $scope.form = {};
+            console.log("Success", resp);
+            window.location.href = "/admin/category/list";
+            window.sessionStorage.setItem("name", "delete");
+          })
+          .catch((error) => {
+            console.log("Error", error);
+            toastMixin.fire({
+              animation: true,
+              icon: "error",
+              title:
+                "Danh mục đã tồn tại trong sản phẩm. Cập nhật không thành công.",
+              position: "top",
+              width: 600,
+            });
+          });
+      }
+    });
   };
   const id = window.sessionStorage.getItem("editId");
   $scope.edit(id);
@@ -204,53 +296,70 @@ function dataFileHandler($scope, $http) {
   //
   // Excel
   $scope.import = (files) => {
-    var confirmImport = confirm("Bạn có muốn import file Excel?");
-    if (confirmImport) {
-      var reader = new FileReader();
-      reader.onloadend = async () => {
-        var workbook = new ExcelJS.Workbook();
-        await workbook.xlsx.load(reader.result);
-        const worksheet = workbook.getWorksheet("category_data");
-        if (!worksheet) {
-          alert("Tên worksheet không đúng. Vui lòng sửa lại tên worksheet.");
-          return;
-        }
-        worksheet.eachRow((row, index) => {
-          if (index > 1) {
-            let student = {
-              id: row.getCell(1).value,
-              name: row.getCell(2).value,
-              description: row.getCell(3).value,
-            };
-            let url = `${host}/category`;
-            $http
-              .post(url, student)
-              .then((resp) => {
-                console.log("Success", resp.data);
-              })
-              .catch((error) => {
-                console.log("Error", error);
-              });
-
-            $scope.load_all();
-          }
+    var reader = new FileReader();
+    reader.onloadend = async () => {
+      var workbook = new ExcelJS.Workbook();
+      await workbook.xlsx.load(reader.result);
+      const worksheet = workbook.getWorksheet("category_data");
+      if (!worksheet) {
+        toastMixin.fire({
+          animation: true,
+          icon: "error",
+          title: "Tên worksheet không đúng. Vui lòng sửa lại tên worksheet.",
+          position: "top",
+          width: 600,
         });
-      };
+        return;
+      }
+      let importSuccess = true;
+      worksheet.eachRow((row, index) => {
+        if (index > 1) {
+          let student = {
+            id: row.getCell(1).value,
+            name: row.getCell(2).value,
+            description: row.getCell(3).value,
+          };
+          let url = `${host}/category`;
+          $http
+            .post(url, student)
+            .then((resp) => {
+              console.log("Success", resp.data);
+            })
+            .catch((error) => {
+              console.log("Error", error);
+              let importSuccess = false;
+            });
 
-      reader.readAsArrayBuffer(files[0]);
-    }
+          $scope.load_all();
+        }
+      });
+      if (importSuccess) {
+        $scope.load_all();
+        toastMixin.fire({
+          animation: true,
+          icon: "success",
+          title: "Import Excel thành công",
+        });
+      } else {
+        toastMixin.fire({
+          animation: true,
+          icon: "error",
+          title: "Import Excel thất bại. Vui lòng kiểm tra lại.",
+        });
+      }
+    };
+
+    reader.readAsArrayBuffer(files[0]);
   };
 
   // Hàm xuất dữ liệu ra tập tin Excel
   $scope.export = () => {
-    var confirmImport = confirm("Bạn có muốn export file Excel?");
-    if (confirmImport) {
     var tableData = [];
-    var headers = ["ID", "FULLNAME", "DESCRIPTION"];
+    var headers = ["ID", "NAME", "DESCRIPTION"];
 
     // Thêm dữ liệu của từng hàng (row) trong bảng vào mảng tableData
     angular.forEach($scope.items, function (item) {
-      var rowData = [item.id, item.name, item.description];
+      var rowData = [item.id, item.name, item.description]; //2
       tableData.push(rowData);
     });
 
@@ -261,12 +370,19 @@ function dataFileHandler($scope, $http) {
     var worksheet = XLSX.utils.aoa_to_sheet([headers].concat(tableData));
 
     // Thêm trang tính vào workbook
-    XLSX.utils.book_append_sheet(workbook, worksheet, "category_data");
+    XLSX.utils.book_append_sheet(workbook, worksheet, "category_data"); // 3
 
     // Xuất file Excel
-    var excelBuffer = XLSX.write(workbook, { bookType: "xlsx", type: "array" });
-    saveAsExcel(excelBuffer, "danh_muc.xlsx");
-  }
+    var excelBuffer = XLSX.write(workbook, {
+      bookType: "xlsx",
+      type: "array",
+    });
+    saveAsExcel(excelBuffer, "category.xlsx");
+    toastMixin.fire({
+      animation: true,
+      icon: "success",
+      title: "Import Excel thành công",
+    });
   };
 
   // Hàm hỗ trợ lưu file Excel
@@ -279,40 +395,98 @@ function dataFileHandler($scope, $http) {
 
   // PDF
   $scope.exportToPDF = function () {
-    var confirmImport = confirm("Bạn có muốn export file PDF?");
-    if (confirmImport) {
     var tableData = [];
-    var headers = ["ID", "Tên danh mục", "Mô tả"];
+    var headers = ["ID", "NAME", "DESCRIPTION"];
 
     // Thêm dữ liệu của từng hàng (row) trong bảng vào mảng tableData
     angular.forEach($scope.items, function (item) {
-      var rowData = [item.id, item.name, item.description];
+      var rowData = [item.id, item.name, item.description]; //2
       tableData.push(rowData);
     });
 
-    //
+    var columnWidths = [];
+
+    var totalColumns = headers.length;
+    var widthPercentage = 100 / totalColumns;
+    for (var i = 0; i < totalColumns; i++) {
+      columnWidths[i] = widthPercentage + "%";
+    }
+
     var docDefinition = {
       content: [
-        { text: "Danh sách danh mục", style: "header" },
+        { text: "Danh sách hãng", style: "header" },
         {
           table: {
             headerRows: 1,
-            widths: ["auto", "*", "*"],
+            widths: columnWidths,
             body: [headers].concat(tableData),
           },
           style: "table",
         },
       ],
-      styles: {
-        header: { fontSize: 20, bold: true, margin: [0, 0, 0, 10] },
-        table: { margin: [0, 5, 0, 15], fontSize: 12 },
-        tableHeader: { fillColor: "#FF0000", bold: true }, // In đậm tiêu đề
-      },
     };
 
+    toastMixin.fire({
+      animation: true,
+      icon: "success",
+      title: "Export PDF thành công",
+    });
     // Xuất PDF
     pdfMake.createPdf(docDefinition).download("danh_muc.pdf");
-  }
   };
   // /PDF
+}
+//
+function validation($scope, item) {
+  var chu =
+    /^[a-zA-Z\sàáạảãâầấậẩẫăằắặẳẵèéẹẻẽêềếệểễìíịỉĩòóọỏõôồốộổỗơờớợởỡùúụủũưừứựửữỳýỵỷỹđÀÁẠẢÃÂẦẤẬẨẪĂẰẮẶẲẴÈÉẸẺẼÊỀẾỆỂỄÌÍỊỈĨÒÓỌỎÕÔỒỐỘỔỖƠỜỚỢỞỠÙÚỤỦŨƯỪỨỰỬỮỲÝỴỶỸĐ]*$/;
+  var kyTuDacBietTen = /[!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?]/;
+  var valid = true;
+
+  if (!item.name) {
+    $scope.MessageName = "Không để rỗng tên danh mục";
+    $scope.showName = true;
+    valid = false;
+  } else if (kyTuDacBietTen.test(item.name)) {
+    $scope.MessageName = "Không được chứa ký tự đặc biệt trong tên danh mục";
+    $scope.showName = true;
+    valid = false;
+  } else if (!chu.test(item.name)) {
+    $scope.MessageName = "Không để số tên danh mục";
+    $scope.showName = true;
+    valid = false;
+  } else {
+    $scope.MessageName = "";
+    $scope.showName = false;
+  }
+
+  if (kyTuDacBietTen.test(item.description)) {
+    $scope.MessageDescription = "Không được chứa ký tự đặc biệt trong mô tả";
+    $scope.showDescription = true;
+    valid = false;
+  } else {
+    $scope.MessageDescription = "";
+    $scope.showDescription = false;
+  }
+
+  return valid;
+}
+function validationCreate(item) {
+  const items = JSON.parse(window.sessionStorage.getItem("items"));
+  var index = items.findIndex(
+    (items) =>
+      items.name.toLowerCase().replace(/\s+/g, "") ===
+      item.name.toLowerCase().replace(/\s+/g, "")
+  );
+  if (index !== -1) {
+    toastMixin.fire({
+      animation: true,
+      icon: "error",
+      title: "Tên danh mục đã tồn tại",
+      position: "top",
+      width: 600,
+    });
+    return false;
+  }
+  return true;
 }
